Adjustments to pebble. (#455)
* Adjustments to pebble.

Remove spawn point for janitor to give a custodial closet [filled]
Removed a part of table in janitors room to give a sink.
Add a sink to kitchen
Adds siobian to mystagogues room
Replaces .45 magnum magazines and ammo boxes to 35 auto for the adjustments of drozds being now the weapon in pebble.

* Gave lawyer's office their own private door

---------

Co-authored-by: Colin-Tel <113523727+Colin-Tel@users.noreply.github.com>